Simplifying the collection view data source into a datasource creator.
